#19 Extract spotify response types from tests into appropriate api modules
