+++ server/src/api/auth/spotify_models.py
@@ -1,0 +1,11 @@
+from typing import TypedDict
+
+from api.common.spotify_models import ImageData
+
+
+class SpotifyFetchMeData(TypedDict):
+    country: str
+    display_name: str
+    id: str
+    images: list[ImageData]
+    product: str
