+++ server/test/pool_features/delete_content_features.py
@@ -1,11 +1,12 @@
 import pytest
+from api.common.spotify_models import TrackData
 from database.database_connection import ConnectionManager
 from database.entities import PoolMember, User
 from helpers.classes import MockedPoolContents
 from sqlalchemy import and_, select
 from starlette.testclient import TestClient
 from test_types.callables import AssertTokenInHeaders, CreatePool, ValidateResponse
-from test_types.typed_dictionaries import Headers, TrackData
+from test_types.typed_dictionaries import Headers
 
 
 def should_delete_track_and_return_remaining_pool_if_given_track_id(
