+++ server/test/pool_features/websocket_features.py
@@ -2,6 +2,7 @@ import datetime
 import random
 
 import pytest
+from api.common.spotify_models import TrackData
 from api.pool import queue_next_songs
 from api.pool.dependencies import PoolPlaybackServiceRaw
 from starlette.testclient import TestClient
@@ -15,7 +16,7 @@ from test_types.callables import (
     SkipSong,
     ValidateResponse,
 )
-from test_types.typed_dictionaries import Headers, TrackData
+from test_types.typed_dictionaries import Headers
 
 
 def should_get_update_when_pool_contents_added(
