+++ server/test/search_features/conftest.py
@@ -2,6 +2,8 @@ import random
 
 import httpx
 import pytest
+from api.common.spotify_models import PaginatedSearchResultData, SpotifyResourceData
+from api.search.spotify_models import GeneralSearchResultData
 from starlette.testclient import TestClient
 from test_types.aliases import MockResponseQueue
 from test_types.callables import (
@@ -16,12 +18,7 @@ from test_types.callables import (
     RunSearchCall,
     ValidatePaginatedResultLength,
 )
-from test_types.typed_dictionaries import (
-    GeneralSearchResultData,
-    Headers,
-    PaginatedSearchResultData,
-    SpotifyResourceData,
-)
+from test_types.typed_dictionaries import Headers
 
 
 @pytest.fixture
