+++ server/test/search_features/general_search_features.py
@@ -3,6 +3,7 @@ from unittest.mock import Mock
 
 import httpx
 import pytest
+from api.common.spotify_models import ImageData
 from conftest import ErrorData
 from starlette.testclient import TestClient
 from test_types.aliases import MockResponseQueue
@@ -18,7 +19,7 @@ from test_types.callables import (
     ValidatePaginatedResultLength,
     ValidateResponse,
 )
-from test_types.typed_dictionaries import Headers, ImageData
+from test_types.typed_dictionaries import Headers
 
 
 @pytest.fixture
